requirements.txt
streamlit
pillow
qrcode
requests